feat: decomposed logic service and configured routing
