+++ src/Components/Router.tsx
@@ -1,0 +1,13 @@
+
+
+const BuildPage = (index: number) => (
+    <>
+        <h3>Page {index}</h3>
+        <div>
+            Page {index} content: { 'asdasdsadsadsadasdasdsad' }
+        </div>
+    </>
+);
+
+export const PageOne = () => BuildPage(1);
+export const PageTwo = () => BuildPage(2);
