+++ src/Pages/Auth/Auth.tsx
@@ -1,0 +1,58 @@
+import {useState} from "react";
+import "./Auth.css"
+
+export default function Auth() {
+    const [isSignUp, setIsSignUp] = useState(true); // Состояние для переключения между формами
+    const toggleForm = () => {setIsSignUp(!isSignUp);};
+
+    return (
+        <div className={`container ${isSignUp ? '' : 'active'}`} id="container">
+            <div className="form-container sign-up">
+                <form>
+                    <h1>Create Account</h1>
+                    <div className="social-icons">
+                        <a href="#" className="icon"><i className="fa-brands fa-google-plus-g"></i></a>
+                        <a href="#" className="icon"><i className="fa-brands fa-facebook-f"></i></a>
+                        <a href="#" className="icon"><i className="fa-brands fa-github"></i></a>
+                        <a href="#" className="icon"><i className="fa-brands fa-linkedin-in"></i></a>
+                    </div>
+                    <span>or use your email for registration</span>
+                    <input type="text" placeholder="Name"/>
+                    <input type="email" placeholder="Email"/>
+                    <input type="password" placeholder="Password"/>
+                    <button type="button">Sign Up</button>
+                </form>
+            </div>
+            <div className="form-container sign-in">
+                <form>
+                    <h1>Sign In</h1>
+                    <div className="social-icons">
+                        <a href="#" className="icon"><i className="fa-brands fa-google-plus-g"></i></a>
+                        <a href="#" className="icon"><i className="fa-brands fa-facebook-f"></i></a>
+                        <a href="#" className="icon"><i className="fa-brands fa-github"></i></a>
+                        <a href="#" className="icon"><i className="fa-brands fa-linkedin-in"></i></a>
+                    </div>
+                    <span>or use your email password</span>
+                    <input type="email" placeholder="Email"/>
+                    <input type="password" placeholder="Password"/>
+                    <a href="#">Forget Your Password?</a>
+                    <button type="button">Sign In</button>
+                </form>
+            </div>
+            <div className="toggle-container">
+                <div className="toggle">
+                    <div className="toggle-panel toggle-left">
+                        <h1>Welcome Back!</h1>
+                        <p>Enter your personal details to use all site features</p>
+                        <button id="login" onClick={toggleForm}>Sign In</button>
+                    </div>
+                    <div className="toggle-panel toggle-right">
+                        <h1>Hello, Friend!</h1>
+                        <p>Register with your personal details to use all site features</p>
+                        <button id="register" onClick={toggleForm}>Sign Up</button>
+                    </div>
+                </div>
+            </div>
+        </div>
+    )
+}
