+++ src/Pages/ChoosGoods/ChooseGoods.tsx
@@ -1,0 +1,7 @@
+export default function ChooseGoods() {
+    return (
+        <>
+            <h1>This page is Ready!</h1>
+        </>
+    )
+}
