+++ vite.config.ts
@@ -3,6 +3,6 @@ import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  base: "/microservice-registration-react/",
+  base: "/",
   plugins: [react()],
 })
